# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /teams:
    post:
      operationId: TeamsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamDto'
      responses:
        '201':
          description: ''
      security:
        - user-auth: []
      summary: Create a new team (Authenticated participant)
      tags:
        - Teams
  /teams/{id}:
    get:
      operationId: TeamsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get team by ID with members and registered events
      tags:
        - Teams
    patch:
      operationId: TeamsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamDto'
      responses:
        '200':
          description: ''
      security:
        - user-auth: []
      summary: Update team name (Team leader only)
      tags:
        - Teams
    delete:
      operationId: TeamsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
        - user-auth: []
      summary: Delete team (Team leader or Admin)
      tags:
        - Teams
  /teams/{id}/leader:
    patch:
      operationId: TeamsController_transferLeadership
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferLeadershipDto'
      responses:
        '200':
          description: ''
      security:
        - user-auth: []
      summary: Transfer team leadership to another team member (Current leader only)
      tags:
        - Teams
  /teams/{id}/join-request:
    post:
      operationId: TeamsController_requestJoin
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinTeamRequestDto'
      responses:
        '201':
          description: ''
      security:
        - user-auth: []
      summary: Request to join a team for an event (Participant)
      tags:
        - Teams
  /teams/{id}/join-requests:
    get:
      operationId: TeamsController_listJoinRequests
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: eventId
          required: false
          in: query
          description: Optional event ID filter for pending requests
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - user-auth: []
      summary: List pending join requests (Team leader only)
      tags:
        - Teams
  /teams/{id}/join-requests/{participantId}:
    patch:
      operationId: TeamsController_reviewJoinRequest
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: participantId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewJoinRequestDto'
      responses:
        '200':
          description: ''
      security:
        - user-auth: []
      summary: Approve or reject a join request (Team leader only)
      tags:
        - Teams
  /teams/{id}/members/{participantId}:
    delete:
      operationId: TeamsController_removeMember
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: participantId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - user-auth: []
      summary: Remove member from team (Team leader only)
      tags:
        - Teams
...
  /members/{id}/positions:
    get:
      operationId: PositionsController_getPositionHistory
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get position history for a member
      tags:
        - Positions
    post:
      operationId: PositionsController_create
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePositionDto'
      responses:
        '201':
          description: ''
      security:
        - team-auth: []
      summary: Assign position to member (Admin only)
      tags:
        - Positions
...
  /positions/year/{year}:
    get:
      description: Returns positions active during the specified year (startYear <= year AND (endYear IS NULL OR endYear >= year))
      operationId: PositionsController_getLeadershipByYear
      parameters:
        - name: year
          required: true
          in: path
          description: Year to query (2000-2100)
          schema:
            type: number
      responses:
        '200':
          description: ''
      summary: Get leadership by year
      tags:
        - Positions
  /positions/{id}:
    patch:
      operationId: PositionsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePositionDto'
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Update position (Admin only)
      tags:
        - Positions
    delete:
      operationId: PositionsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Remove position (Admin only)
      tags:
        - Positions
  /projects:
    get:
      operationId: ProjectsController_findAll
      parameters:
        - name: limit
          required: false
          in: query
          description: Number of items to return
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: offset
          required: false
          in: query
          description: Number of items to skip
          schema:
            minimum: 0
            default: 0
            type: number
        - name: search
          required: false
          in: query
          description: Search by project title
          schema:
            type: string
        - name: tags
          required: false
          in: query
          description: Filter by tags
          schema:
            type: array
            items:
              type: string
        - name: memberId
          required: false
          in: query
          description: Filter by member ID
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: List all projects with filters
      tags:
        - Projects
    post:
      operationId: ProjectsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectDto'
      responses:
        '201':
          description: ''
      security:
        - team-auth: []
      summary: Create a new project (Member only)
      tags:
        - Projects
  /projects/{id}:
    get:
      operationId: ProjectsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get project by ID with members
      tags:
        - Projects
    patch:
      operationId: ProjectsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectDto'
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Update project (Project member only)
      tags:
        - Projects
    delete:
      operationId: ProjectsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Delete project (Project creator or Admin only)
      tags:
        - Projects
  /projects/{id}/members:
    post:
      operationId: ProjectsController_addMember
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectMemberDto'
      responses:
        '201':
          description: ''
      security:
        - team-auth: []
      summary: Add member to project (Project member only)
      tags:
        - Projects
  /projects/{projectId}/members/{memberId}:
    delete:
      operationId: ProjectsController_removeMember
      parameters:
        - name: projectId
          required: true
          in: path
          schema:
            type: string
        - name: memberId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Remove member from project (Project member only)
      tags:
        - Projects
...
info:
  title: The Robotics Club API
  description: API for the college robotics club management platform
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local Development
  - url: https://api.roboticsclub.example.com
    description: Production
components:
  securitySchemes:
    user-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token for participants (external users)
    team-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token for club members/admins
  schemas:
...
    CreateTeamDto:
      type: object
      properties:
        name:
          type: string
          description: Team name
          example: Robo Warriors
      required:
        - name
    UpdateTeamDto:
      type: object
      properties:
        name:
          type: string
          description: Team name
    TransferLeadershipDto:
      type: object
      properties:
        newLeaderId:
          type: string
          description: Participant ID to transfer leadership to
      required:
        - newLeaderId
    JoinTeamRequestDto:
      type: object
      properties:
        eventId:
          type: string
          description: Event ID for this join request
      required:
        - eventId
    ReviewJoinRequestDto:
      type: object
      properties:
        eventId:
          type: string
          description: Event ID for this join request
        action:
          type: string
          description: Decision for the participant join request
          enum:
            - APPROVE
            - REJECT
      required:
        - eventId
        - action
...
    CreatePositionDto:
      type: object
      properties:
        position:
          type: string
          description: Position title
          enum:
            - CHAIRPERSON
            - VICE_CHAIRPERSON
            - TREASURER
            - TECHNICAL_HEAD
            - CODING_HEAD
            - LOGISTICS_HEAD
            - EXECUTIVE_MEMBER
            - MEMBER
        startMonth:
          type: number
          description: Start month (1-12)
          example: 1
          minimum: 1
          maximum: 12
        startYear:
          type: number
          description: Start year
          example: 2024
          minimum: 2000
          maximum: 2100
        endMonth:
          type: number
          description: End month (1-12, null for current)
          minimum: 1
          maximum: 12
        endYear:
          type: number
          description: End year (null for current)
          minimum: 2000
          maximum: 2100
      required:
        - position
        - startMonth
        - startYear
    UpdatePositionDto:
      type: object
      properties:
        position:
          type: string
          description: Position title
          enum:
            - CHAIRPERSON
            - VICE_CHAIRPERSON
            - TREASURER
            - TECHNICAL_HEAD
            - CODING_HEAD
            - LOGISTICS_HEAD
            - EXECUTIVE_MEMBER
            - MEMBER
        startMonth:
          type: number
          description: Start month (1-12)
          minimum: 1
          maximum: 12
        startYear:
          type: number
          description: Start year
          minimum: 2000
          maximum: 2100
        endMonth:
          type: number
          description: End month (1-12)
          minimum: 1
          maximum: 12
        endYear:
          type: number
          description: End year
          minimum: 2000
          maximum: 2100
    CreateProjectDto:
      type: object
      properties:
        title:
          type: string
          description: Project title
          example: Autonomous Robot
        description:
          type: string
          description: Short project description
        content:
          type: object
          description: Rich content (Quill Delta JSON)
        contentHtml:
          type: string
          description: HTML representation of content
        imageUrl:
          type: string
          description: Project image URL
        githubRepo:
          type: string
          description: GitHub repository URL
        demoUrl:
          type: string
          description: Demo/live URL
        tags:
          description: Project tags
          type: array
          items:
            type: string
      required:
        - title
        - description
        - content
        - contentHtml
    UpdateProjectDto:
      type: object
      properties:
        title:
          type: string
          description: Project title
        description:
          type: string
          description: Project description
        content:
          type: object
          description: Rich content (Quill Delta JSON)
        contentHtml:
          type: string
          description: HTML representation of content
        imageUrl:
          type: string
          description: Project image URL
        githubRepo:
          type: string
          description: GitHub repository URL
        demoUrl:
          type: string
          description: Demo/live URL
        tags:
          description: Project tags
          type: array
          items:
            type: string
    AddProjectMemberDto:
      type: object
      properties:
        memberId:
          type: string
          description: Member ID to add to project
        role:
          type: string
          description: Role in project
          example: Lead Developer
      required:
        - memberId
...